HDDS-11852. Reduce duplication in some GenericCli subclasses
