+++ hadoop-ozone/insight/src/main/java/org/apache/hadoop/ozone/insight/Insight.java
@@ -34,10 +34,6 @@ import picocli.CommandLine;
     mixinStandardHelpOptions = true)
 public class Insight extends GenericCli {
 
-  public Insight() {
-    super(Insight.class);
-  }
-
   public static void main(String[] args) throws Exception {
     new Insight().run(args);
   }
