+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/repair/OzoneRepair.java
@@ -18,12 +18,10 @@
 
 package org.apache.hadoop.ozone.repair;
 
-import com.google.common.annotations.VisibleForTesting;
 import org.apache.hadoop.hdds.cli.ExtensibleParentCommand;
 import org.apache.hadoop.hdds.cli.GenericCli;
 import org.apache.hadoop.hdds.cli.HddsVersionProvider;
 import org.apache.hadoop.hdds.cli.RepairSubcommand;
-import org.apache.hadoop.hdds.conf.OzoneConfiguration;
 import picocli.CommandLine;
 
 import java.nio.charset.StandardCharsets;
@@ -42,33 +40,7 @@ public class OzoneRepair extends GenericCli implements ExtensibleParentCommand {
       "ATTENTION: Running as user %s. Make sure this is the same user used to run the Ozone process." +
           " Are you sure you want to continue (y/N)? ";
 
-
-  private OzoneConfiguration ozoneConf;
-
-  public OzoneRepair() {
-    super(OzoneRepair.class);
-  }
-
-  @VisibleForTesting
-  public OzoneRepair(OzoneConfiguration configuration) {
-    super(OzoneRepair.class);
-    this.ozoneConf = configuration;
-  }
-
-  public OzoneConfiguration getOzoneConf() {
-    if (ozoneConf == null) {
-      ozoneConf = createOzoneConfiguration();
-    }
-    return ozoneConf;
-  }
-
-  /**
-   * Main for the Ozone Repair shell Command handling.
-   *
-   * @param argv - System Args Strings[]
-   * @throws Exception
-   */
-  public static void main(String[] argv) throws Exception {
+  public static void main(String[] argv) {
     new OzoneRepair().run(argv);
   }
